+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.5 (unreleased)
 ================
 
+- registry.regsearch now accepts an optional maxrec argument rather than
+  automatically passing the service's hard limit.
+
 
 1.4 (2022-09-26)
 ================