+++ CHANGES.rst
@@ -20,6 +20,10 @@
 
 - Registry search now finds SIA v2 services. [#422]
 
+- Error messages from uws jobs are now in job.errorsummary.message
+  rather than job.message (where one wouldn't expect them given the UWS
+  schema). [#432]
+
 
 1.4 (2022-09-26)
 ================