+++ CHANGES.rst
@@ -4,12 +4,15 @@
 - registry.regsearch now accepts an optional maxrec argument rather than
   automatically passing the service's hard limit. [#375]
 
+- fixed the RegTAP fragment for the discovery of EPN-TAP tables. [#395]
+
 
 1.4 (2022-09-26)
 ================
+
 - Added the TAP Table Manipulation prototype (cadc-tb-upload). [#274]
 
-- More explicit exception messages where the payload is 
+- More explicit exception messages where the payload is
   sometimes considered if it can be presented properly (simple
   body text or job error message). [#355]